Do not support SSH out of the box
As SSH needs priv/pub keys we would needed to mount a static key for each of the images
which in the end did not seem like a good idea for usability or security standpoint.
The main point of having SSH support was to install dependencies over SSH, but since
dependencies can be installed over HTTPS as well, that is the recommended way from
now on.