Explicitly set appuser to have uid/gid 1000
While this is not perfect and will not work in every case, it is a good default
value. Users might have different ids on their host machine which might not
correspond to this value, in those cases the current solution is to simply add
set the host folder to have the owner 1000. Again, not a perfect solution, but
a good default workaround.

This could be further extended by using tools for correcting the uid/gid in the
image. Those often requires that the user would not be created in the base image
however, but instead in the application image. Currently we are specifying the
appuser in the base image however for the reason that it makes the development
work easier for developers so that they do not need to create a user as part of
them creating a Dockerimage. The other way around this issue to use a run-time
wrapper around docker-compose which changes the user gid/uid that runs the
application image. This is something that can be considered in the future if this
becomes a larger issue.